+++ comic-safe/package.json
@@ -13,6 +13,7 @@
     "@nuxt/eslint": "^1.4.1",
     "@nuxt/fonts": "^0.11.4",
     "@nuxt/ui": "^3.1.2",
+    "cheerio": "^1.0.0",
     "eslint": "^9.27.0",
     "nuxt": "^3.17.4",
     "typescript": "^5.8.3",
